added --keep-tmp and --model options

+++ influenza_consensus.sh
@@ -11,8 +11,10 @@ Required arguments:
 
 Optional arguments:
 -t|--threads        Number of threads [Default = 32]
--s|--segment        Limit consensus sequence calling for specific Influenza A genomic segments with each segment number delimited by a comma (Example: -s 1,2,5,6)
+-s|--segment        Target specific Influenza A genomic segments for consensus calling with each segment number delimited by a comma (Example: -s 1,2,5,6)
+-m|--model          Specify the flowcell chemistry used for Nanopore sequencing {Options: r9, r10} [Default = r9]
 --notrim            Disable adaptor trimming by Porechop
+--keep-tmp          Keep all temporary files
 -h|--help           Display help message
 "
 }
@@ -21,7 +23,7 @@ Optional arguments:
 script_dir=$(dirname $(realpath $0))
 
 # parse arguments
-opts=`getopt -o hi:o:t:s: -l help,input:,output:,threads:,db:,notrim,segment: -- "$@"`
+opts=`getopt -o hi:o:t:s:m: -l help,input:,output:,threads:,db:,notrim,segment:,model:,keep-tmp -- "$@"`
 eval set -- "$opts"
 if [ $? != 0 ] ; then echo "influenza_consensus: Invalid arguments used, exiting"; usage; exit 1 ; fi
 if [[ $1 =~ ^--$ ]] ; then echo "influenza_consensus: Invalid arguments used, exiting"; usage; exit 1 ; fi
@@ -32,7 +34,9 @@ while true; do
         -o|--output) OUTPUT_PATH=$2; shift 2;;
         --db) DB_PATH=$2; shift 2;;
         -t|--threads) THREADS=$2; shift 2;;
+        -m|--model) MODEL=$2; shift 2;;
         --notrim) TRIM=0; shift 1;;
+        --keep-tmp) KEEP_TMP=1; shift 1;;
         -s|--segment) SEGMENTS=$2; shift 2;;
         --) shift; break ;;
         -h|--help) usage; exit 0;;
@@ -62,18 +66,37 @@ if [[ $? != 0 ]]; then echo "influenza_consensus: centrifuge cannot be called, c
 porechop -h > /dev/null
 if [[ $? != 0 ]]; then echo "influenza_consensus: porechop cannot be called, check its installation"; exit 1; fi
 
-# validate segment input
-for segment in $(echo $SEGMENTS | sed 's/,/\n/g'); do
-  # check if individual elements are valid integers between [1:8]
-  if [[ $(echo -n $segment | wc -m) -ne 1 ]]; then
-    echo "Invalid characters passed to the -s argument, exiting"
-    exit 1
-  fi
-  if ! [[ $segment =~ ^[1-8]$ ]]; then
-    echo "Invalid characters passed to the -s argument, exiting"
-    exit 1
-  fi
-done
+NanoPlot -h > /dev/null
+if [[ $? != 0 ]]; then echo "influenza_consensus: nanoplot cannot be called, check its installation"; exit 1; fi
+
+# validate model parameter input if specified
+if ! test -z $MODEL; then
+  # test if invalid characters used
+  if ! [[ $MODEL =~ ^(r9|r10)$ ]]; then echo "Invalid model specification passed to the -m argument, exiting"; fi
+  # set medaka model
+  if [[ $MODEL == "r9" ]]; then MODEL="r941_min_high_g360"; else MODEL="r103_min_high_g360"; fi
+else
+  # Set default model if not specified
+  if test -z $MODEL; then MODEL="r941_min_high_g360"; fi
+fi
+
+# validate segment parameter input if specified
+if ! test -z $SEGMENTS; then
+  for segment in $(echo $SEGMENTS | sed 's/,/\n/g'); do
+    # check if individual elements are valid integers between [1:8]
+    if [[ $(echo -n $segment | wc -m) -ne 1 ]]; then
+      echo "Invalid characters passed to the -s argument, exiting"
+      exit 1
+    fi
+    if ! [[ $segment =~ ^[1-8]$ ]]; then
+      echo "Invalid characters passed to the -s argument, exiting"
+      exit 1
+    fi
+  done
+else
+  # Set default segments to produce = [1:8] if not specified
+  if test -z $SEGMENTS; then SEGMENTS="1,2,3,4,5,6,7,8,"; fi
+fi
 
 # validate input samples.csv
 if ! test -f $INPUT_PATH; then echo "influenza_consensus: Input sample file does not exist, exiting"; exit 1; fi
@@ -99,24 +122,28 @@ if test -z $THREADS; then THREADS=32; fi
 # Set default trim mode to true if not specified
 if test -z $TRIM; then TRIM=1; fi
 
-# Set default segments to produce to [1:8] if not specified
-if test -z $SEGMENTS; then SEGMENTS="1,2,3,4,5,6,7,8,"; fi
+# Set default keep temporary files (KEEP_TMP) to 0 if not specified
+if test -z $KEEP_TMP; then TMP=0; fi
 
 # call snakemake
-snakemake --dag --snakefile $script_dir/SnakeFile --cores $THREADS \
+snakemake --snakefile $script_dir/SnakeFile --cores $THREADS \
   --config samples=$(realpath $INPUT_PATH) \
-  outdir=$OUTPUT_PATH \
+  outdir=$(realpath $OUTPUT_PATH) \
   segments="$SEGMENTS" \
   pipeline_dir=$script_dir \
   centrifuge_db=$DB_PATH \
-  trim=$TRIM
+  trim=$TRIM \
+  model=$MODEL \
+  threads=$THREADS
 
 # clean up temporary directories
-while read lines; do
-  sample=$(echo $lines | cut -f1 -d',')
-  for dir in fastq medaka centrifuge target racon porechop; do
-    if test -d $OUTPUT_PATH/$sample/$dir; then
-      rm -rf $OUTPUT_PATH/$sample/$dir
-    fi
-  done
-done < $INPUT_PATH
+if [[ $KEEP_TMP -eq 0 ]]; then
+  while read lines; do
+    sample=$(echo $lines | cut -f1 -d',')
+    for dir in nanoplot fastq medaka centrifuge target racon porechop; do
+      if test -d $OUTPUT_PATH/$sample/$dir; then
+        rm -rf $OUTPUT_PATH/$sample/$dir
+      fi
+    done
+  done < $INPUT_PATH
+fi
